server/requirements.txt
flask==3.1.0
flask-cors==5.0.1
supabase==2.13.0
python-dotenv==1.0.1
pyyaml==6.0.2
pyjwt==2.10.1
pytest==8.3.4
requests==2.32.3
